requirements.txt
# Flask core
Flask==3.0.0
flask-cors==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Validation
pydantic==2.5.0

# Environment
python-dotenv==1.0.0

# Scheduler
APScheduler==3.10.4

# JWT tokens
PyJWT==2.8.0

# Email
Jinja2==3.1.2

# HTTP client
requests==2.31.0
